+++ connectome_tools/s2f_recipe/estimate_individual_bouton_reduction.py
@@ -16,13 +16,14 @@ from bluepy import Cell, Circuit
 
 from connectome_tools.dataset import read_bouton_density
 from connectome_tools.s2f_recipe import BOUTON_REDUCTION_FACTOR
-from connectome_tools.s2f_recipe.utils import Task
+from connectome_tools.s2f_recipe.utils import BaseExecutor
 from connectome_tools.stats import sample_bouton_density
+from connectome_tools.utils import Task
 
 L = logging.getLogger(__name__)
 
 
-def estimate_bouton_density(
+def _estimate_bouton_density(
     circuit_config, mtype, sample_size, sample_target, mask, assume_syns_bouton
 ):
     """Mean bouton density for given mtype."""
@@ -36,36 +37,53 @@ def estimate_bouton_density(
     return np.nanmean(values)
 
 
-def prepare(circuit, bio_data, sample=None):
-    # noqa: D103 # pylint: disable=missing-docstring
-    mtypes = circuit.cells.mtypes
-    if isinstance(bio_data, float):
-        bio_data = pd.DataFrame(
-            {
-                "mtype": mtypes,
-                "mean": bio_data,
-            }
-        )
-    else:
-        bio_data = read_bouton_density(bio_data, mtypes=mtypes)
+class Executor(BaseExecutor):
+    """Executor class for estimate_individual_bouton_reduction strategy."""
 
-    if isinstance(sample, str):
-        dset = read_bouton_density(sample).set_index("mtype")
-        estimate = lambda mtype: dset.loc[mtype]["mean"]
-    else:
-        if sample is None:
-            sample = {}
-        estimate = partial(
-            estimate_bouton_density,
-            circuit_config=circuit.config,
-            sample_size=sample.get("size", 100),
-            sample_target=sample.get("target", None),
-            mask=sample.get("mask", None),
-            assume_syns_bouton=sample.get("assume_syns_bouton", 1.0),
-        )
+    is_parallel = True
 
-    for _, row in bio_data.iterrows():
-        yield Task(_execute, row, estimate, task_group=__name__)
+    def prepare(self, circuit, bio_data, sample=None):
+        """Yield tasks that should be executed.
+
+        Args:
+            circuit (bluepy.Circuit): circuit instance.
+            bio_data: reference value (float)
+                or name of the .tsv file containing bouton density data (str).
+            sample: sample configuration (dict)
+                or name of the .tsv file containing bouton density data (str).
+
+        Yields:
+            (Task) task to be executed.
+        """
+        # pylint: disable=arguments-differ
+        mtypes = circuit.cells.mtypes
+        if isinstance(bio_data, float):
+            bio_data = pd.DataFrame(
+                {
+                    "mtype": mtypes,
+                    "mean": bio_data,
+                }
+            )
+        else:
+            bio_data = read_bouton_density(bio_data, mtypes=mtypes)
+
+        if isinstance(sample, str):
+            dset = read_bouton_density(sample).set_index("mtype")
+            estimate = lambda mtype: dset.loc[mtype]["mean"]
+        else:
+            if sample is None:
+                sample = {}
+            estimate = partial(
+                _estimate_bouton_density,
+                circuit_config=circuit.config,
+                sample_size=sample.get("size", 100),
+                sample_target=sample.get("target", None),
+                mask=sample.get("mask", None),
+                assume_syns_bouton=sample.get("assume_syns_bouton", 1.0),
+            )
+
+        for _, row in bio_data.iterrows():
+            yield Task(_execute, row, estimate, task_group=__name__)
 
 
 def _execute(row, estimate):
@@ -74,5 +92,5 @@ def _execute(row, estimate):
     if np.isnan(value):
         L.warning("Could not estimate '%s' bouton density, skipping", mtype)
         return []
-    L.debug("Bouton density estimate for '%s': %.3g", mtype, value)
+    L.info("Bouton density estimate for '%s': %.3g", mtype, value)
     return [((mtype, "*"), {BOUTON_REDUCTION_FACTOR: ref_value / value})]