+++ connectome_tools/s2f_recipe/experimental_syns_con.py
@@ -7,12 +7,27 @@ has been experimentally determined.
 
 from connectome_tools.dataset import read_nsyn
 from connectome_tools.s2f_recipe import MEAN_SYNS_CONNECTION
-from connectome_tools.s2f_recipe.utils import Task
+from connectome_tools.s2f_recipe.utils import BaseExecutor
+from connectome_tools.utils import Task
 
 
-def prepare(circuit, bio_data):
-    # noqa: D103 # pylint: disable=missing-docstring
-    yield Task(_execute, bio_data, mtypes=circuit.cells.mtypes, task_group=__name__)
+class Executor(BaseExecutor):
+    """Executor class for experimental_syns_con strategy."""
+
+    is_parallel = False
+
+    def prepare(self, circuit, bio_data):
+        """Yield tasks that should be executed.
+
+        Args:
+            circuit (bluepy.Circuit): circuit instance.
+            bio_data (str): name of the .tsv file containing experimental nsyn data.
+
+        Yields:
+            (Task) task to be executed.
+        """
+        # pylint: disable=arguments-differ
+        yield Task(_execute, bio_data, mtypes=circuit.cells.mtypes, task_group=__name__)
 
 
 def _execute(bio_data, mtypes):