+++ tests/s2f_recipe/test_estimate_individual_bouton_reduction.py
@@ -30,7 +30,7 @@ def test_1(_):
         ("L23_MC", "*"): {"bouton_reduction_factor": 5.0},
         ("L5_TPC", "*"): {"bouton_reduction_factor": 10.0},
     }
-    task_generator = test_module.prepare(circuit, bio_data=10.0)
+    task_generator = test_module.Executor().prepare(circuit, bio_data=10.0)
     result_generator = (task() for task in task_generator)
     actual = dict(chain.from_iterable(item.value for item in result_generator))
 
@@ -46,7 +46,7 @@ def test_2(_):
         ("L23_MC", "*"): {"bouton_reduction_factor": 15.0},
     }
     bio_data = os.path.join(TEST_DATA_DIR, "bouton_density.tsv")
-    task_generator = test_module.prepare(circuit, bio_data=bio_data)
+    task_generator = test_module.Executor().prepare(circuit, bio_data=bio_data)
     result_generator = (task() for task in task_generator)
     actual = dict(chain.from_iterable(item.value for item in result_generator))