+++ tests/s2f_recipe/test_generalized_cv.py
@@ -11,7 +11,7 @@ def test_prepare():
     expected = {("*", "*"): {"cv_syns_connection": cv}}
     circuit = MagicMock(Circuit)
 
-    task_generator = test_module.prepare(circuit, cv)
+    task_generator = test_module.Executor().prepare(circuit, cv)
     result_generator = (task() for task in task_generator)
     actual = dict(chain.from_iterable(item.value for item in result_generator))