+++ tests/s2f_recipe/test_override_mtype.py
@@ -57,7 +57,7 @@ def test_prepare(_, kwargs, expected):
     circuit = MagicMock(Circuit)
     circuit.cells.mtypes = {"L6_TPC:C", "L4_CHC"}
 
-    task_generator = test_module.prepare(circuit, **kwargs)
+    task_generator = test_module.Executor().prepare(circuit, **kwargs)
     result_generator = (task() for task in task_generator)
     actual = dict(chain.from_iterable(item.value for item in result_generator))